all: Implement log streaming (#4)
Add a `telejob logs` command to retrieve and continuously print logs of a
running job or retrieve historical logs of a stopped job.

Implement the streaming `Logs` method of the gRPC service to stream log data to
the client CLI.

The service method calls the job controller's `Logs` method, which creates a
private `logDispatcher` for each job. The `logDispatcher` retrieves `Stdout`
and `Stderr` from the job's process and dispatches this output to all readers,
ultimately sending it on the gRPC server stream.

Write tests, stress tests, and documentation for the new feature.

Pull-request: https://github.com/juliaogris/telejob/pull/4